+++ core/interfaces/core.go
@@ -34,6 +34,12 @@ type Client interface {
 	CreateUserMatchingResult(example model.UserMatchingResult) (*model.UserMatchingResult, error)
 	UpdateUserMatchingResult(example model.UserMatchingResult) error
 	DeleteUserMatchingResult(id string) error
+
+	// Survey Data APIs
+	GetSurveyData(id string) (*model.SurveyData, error)
+	CreateSurveyData(example model.SurveyData) (*model.SurveyData, error)
+	UpdateSurveyData(example model.SurveyData) error
+	DeleteSurveyData(id string) error
 }
 
 // Admin exposes administrative APIs for the driver adapters