+++ core/interfaces/mocks/Storage.go
@@ -22,6 +22,20 @@ func (_m *Storage) CreateUserMatchingResult(bessiData model.UserMatchingResult) 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(model.UserMatchingResult) error); ok {
 		r0 = rf(bessiData)
+  } else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// CreateSurveyData provides a mock function with given fields: surveyData
+func (_m *Storage) CreateSurveyData(surveyData model.SurveyData) error {
+	ret := _m.Called(surveyData)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(model.SurveyData) error); ok {
+		r0 = rf(surveyData)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -59,6 +73,20 @@ func (_m *Storage) DeleteExample(orgID string, appID string, id string) error {
 
 // DeleteUserMatchingResult provides a mock function with given fields: id
 func (_m *Storage) DeleteUserMatchingResult(id string) error {
+	ret := _m.Called(id)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(string) error); ok {
+		r0 = rf(id)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// DeleteSurveyData provides a mock function with given fields: id
+func (_m *Storage) DeleteSurveyData(id string) error {
 	ret := _m.Called(id)
 
 	var r0 error
@@ -201,6 +229,32 @@ func (_m *Storage) GetUserMatchingResult(id string) (*model.UserMatchingResult, 
 	return r0, r1
 }
 
+// GetSurveyData provides a mock function with given fields: id
+func (_m *Storage) GetSurveyData(id string) (*model.SurveyData, error) {
+	ret := _m.Called(id)
+
+	var r0 *model.SurveyData
+	var r1 error
+	if rf, ok := ret.Get(0).(func(string) (*model.SurveyData, error)); ok {
+		return rf(id)
+	}
+	if rf, ok := ret.Get(0).(func(string) *model.SurveyData); ok {
+		r0 = rf(id)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*model.SurveyData)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(string) error); ok {
+		r1 = rf(id)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 // InsertConfig provides a mock function with given fields: config
 func (_m *Storage) InsertConfig(config model.Config) error {
 	ret := _m.Called(config)
@@ -283,6 +337,20 @@ func (_m *Storage) UpdateUserMatchingResult(bessiData model.UserMatchingResult) 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(model.UserMatchingResult) error); ok {
 		r0 = rf(bessiData)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
+// UpdateSurveyData provides a mock function with given fields: surveyData
+func (_m *Storage) UpdateSurveyData(surveyData model.SurveyData) error {
+	ret := _m.Called(surveyData)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(model.SurveyData) error); ok {
+		r0 = rf(surveyData)
 	} else {
 		r0 = ret.Error(0)
 	}