+++ driver/web/apis_client.go
@@ -38,12 +38,12 @@ func (h ClientAPIsHandler) getExample(l *logs.Log, r *http.Request, claims *toke
 		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
 	}
 
-	example, err := h.app.Client.GetExample(claims.OrgID, claims.AppID, id)
+	surveyData, err := h.app.Client.GetExample(claims.OrgID, claims.AppID, id)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeExample, nil, err, http.StatusInternalServerError, true)
 	}
 
-	response, err := json.Marshal(example)
+	response, err := json.Marshal(surveyData)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponseBody, nil, err, http.StatusInternalServerError, false)
 	}
@@ -56,13 +56,32 @@ func (h ClientAPIsHandler) getUserMatchingResult(l *logs.Log, r *http.Request, c
 	if len(id) <= 0 {
 		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
 	}
-
-	userMatchingResult, err := h.app.Client.GetUserMatchingResult(id)
+  
+  userMatchingResult, err := h.app.Client.GetUserMatchingResult(id)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
 	}
 
 	response, err := json.Marshal(userMatchingResult)
+  if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponseBody, nil, err, http.StatusInternalServerError, false)
+	}
+	return l.HTTPResponseSuccessJSON(response)
+}
+
+func (h ClientAPIsHandler) getSurveyData(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
+	params := mux.Vars(r)
+	id := params["id"]
+	if len(id) <= 0 {
+		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
+	}
+
+	surveyData, err := h.app.Client.GetSurveyData(id)
+	if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeSurveyData, nil, err, http.StatusInternalServerError, true)
+	}
+
+	response, err := json.Marshal(surveyData)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponseBody, nil, err, http.StatusInternalServerError, false)
 	}
@@ -71,11 +90,11 @@ func (h ClientAPIsHandler) getUserMatchingResult(l *logs.Log, r *http.Request, c
 
 func (h ClientAPIsHandler) createUserMatchingResult(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
 	var requestData model.UserMatchingResult
-	err := json.NewDecoder(r.Body).Decode(&requestData)
+  err := json.NewDecoder(r.Body).Decode(&requestData)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
 	}
-
+  
 	userMatchingResult, err := h.app.Client.CreateUserMatchingResult(requestData)
 	if err != nil || userMatchingResult == nil {
 		return l.HTTPResponseErrorAction(logutils.ActionCreate, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
@@ -88,23 +107,65 @@ func (h ClientAPIsHandler) createUserMatchingResult(l *logs.Log, r *http.Request
 	return l.HTTPResponseSuccessJSON(response)
 }
 
+func (h ClientAPIsHandler) createSurveyData(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
+	var requestData model.SurveyData
+	err := json.NewDecoder(r.Body).Decode(&requestData)
+	if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
+	}
+  
+	surveyData, err := h.app.Client.CreateSurveyData(requestData)
+	surveyData.ID = claims.Subject
+	if err != nil || surveyData == nil {
+		return l.HTTPResponseErrorAction(logutils.ActionCreate, model.TypeSurveyData, nil, err, http.StatusInternalServerError, true)
+	}
+
+	response, err := json.Marshal(surveyData)
+	if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponseBody, nil, err, http.StatusInternalServerError, false)
+	}
+	return l.HTTPResponseSuccessJSON(response)
+}
+
 func (h ClientAPIsHandler) updateUserMatchingResult(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
 	params := mux.Vars(r)
 	id := params["id"]
 	if len(id) <= 0 {
 		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
 	}
+  
+  var requestData model.UserMatchingResult
+  err := json.NewDecoder(r.Body).Decode(&requestData)
+	if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
+	}
 
-	var requestData model.UserMatchingResult
+	requestData.ID = id
+  err = h.app.Client.UpdateUserMatchingResult(requestData)
+	if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionUpdate, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
+	}
+
+	return l.HTTPResponseSuccess()
+}
+
+func (h ClientAPIsHandler) updateSurveyData(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
+	params := mux.Vars(r)
+	id := params["id"]
+	if len(id) <= 0 {
+		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
+	}
+
+	var requestData model.SurveyData
 	err := json.NewDecoder(r.Body).Decode(&requestData)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
 	}
 
 	requestData.ID = id
-	err = h.app.Client.UpdateUserMatchingResult(requestData)
+	err = h.app.Client.UpdateSurveyData(requestData)
 	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionUpdate, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
+		return l.HTTPResponseErrorAction(logutils.ActionUpdate, model.TypeSurveyData, nil, err, http.StatusInternalServerError, true)
 	}
 
 	return l.HTTPResponseSuccess()
@@ -120,6 +181,21 @@ func (h ClientAPIsHandler) deleteUserMatchingResult(l *logs.Log, r *http.Request
 	err := h.app.Client.DeleteUserMatchingResult(id)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionDelete, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
+  }
+
+	return l.HTTPResponseSuccess()
+}
+
+func (h ClientAPIsHandler) deleteSurveyData(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
+	params := mux.Vars(r)
+	id := params["id"]
+	if len(id) <= 0 {
+		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
+	}
+
+	err := h.app.Client.DeleteSurveyData(id)
+	if err != nil {
+		return l.HTTPResponseErrorAction(logutils.ActionDelete, model.TypeSurveyData, nil, err, http.StatusInternalServerError, true)
 	}
 
 	return l.HTTPResponseSuccess()